requirements.txt
blinker==1.6.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
docopt==0.6.2
Flask==2.3.2
Flask-Minify==0.42
gunicorn==21.2.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lesscpy==0.15.1
MarkupSafe==2.1.2
packaging==23.2
pipreqs==0.4.13
ply==3.11
PyYAML==6.0
rcssmin==1.1.2
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==2.3.4
xxhash==3.4.1
yarg==0.1.9
zipp==3.15.0
